pyproject.toml
[tool.poetry]
name = "poc-fastapi"
version = "0.1.0"
description = ""
authors = ["Patrick Rodrigues <patrick.pwall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
SQLAlchemy = "^1.4.17"
pyhocon = "^0.3.57"
asyncpg = "^0.23.0"
mixer = "^7.1.2"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = {version = "^3.2.0", extras = ["cryptography"]}
httpx = "^0.18.1"
SQLAlchemy-Utils = "^0.37.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.5b2"
isort = "^5.8.0"
flake8 = "^3.9.2"
mypy = "^0.812"
pytest-cov = "^2.12.0"
pre-commit = "^2.13.0"
pytest-env = "^0.6.2"
pytest-mock = "^3.6.1"
pytest-asyncio = "^0.15.1"
aiosqlite = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
